+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 * `AnimationState::from_frame_duration_secs`
+* `AnimationState::update_secs`
 
 
 ## [0.1.0] - 2023-12-10